requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pandas==2.2.2
scikit-learn==1.4.2
python-multipart==0.0.9
python-dotenv==1.0.1
apscheduler==3.10.4
pytest==8.2.0
httpx==0.27.0
pydantic-settings==2.2.1
